Revert "sched: Fix sched_policy < 0 comparison"
This reverts commit e73519117009dfe94bbcf499104488ca8db7b3cc.

+++ kernel/sched/core.c
@@ -6406,7 +6406,7 @@ SYSCALL_DEFINE3(sched_setattr, pid_t, pid, struct sched_attr __user *, uattr,
 	if (retval)
 		return retval;
 
-	if ((int)attr.sched_policy < 0)
+	if (attr.sched_policy < 0)
 		return -EINVAL;
 
 	rcu_read_lock();
